Backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-extensions==4.1
django-tenants==3.5.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
pandas==2.2.3
openpyxl==3.1.2
